Share translations with PicoArch
