+++ lang/builtin/en_US.json
@@ -80,4 +80,36 @@
   ,"app.videos": "Videos"
   ,"app.music_mp3": "Music / MP3"
   ,"app.rockbox": "Rockbox"
+  ,"pico.menu.resume": "Resume game"
+  ,"pico.menu.save_state": "Save state"
+  ,"pico.menu.load_state": "Load state"
+  ,"pico.menu.disc_control": "Disc control"
+  ,"pico.menu.cheats": "Cheats"
+  ,"pico.menu.options": "Options"
+  ,"pico.menu.reset_game": "Reset game"
+  ,"pico.menu.load_new_game": "Load new game"
+  ,"pico.menu.exit": "Exit"
+  ,"pico.options.audio_video": "Audio and video"
+  ,"pico.options.emulator": "Emulator options"
+  ,"pico.options.player_controls": "Player controls"
+  ,"pico.options.emulator_controls": "Emulator controls"
+  ,"pico.options.save_config": "Save config"
+  ,"pico.video.show_fps": "Show FPS"
+  ,"pico.video.show_cpu": "Show CPU %"
+  ,"pico.video.fast_forward": "Fast forward (SELECT+R1)"
+  ,"pico.video.rewind": "Rewind (hold SELECT+B)"
+  ,"pico.video.aspect_ratio": "Aspect ratio"
+  ,"pico.video.filter": "Filter"
+  ,"pico.video.optimize_text": "Optimize text"
+  ,"pico.video.audio_buffer": "Audio buffer"
+  ,"pico.video.mono_output": "Mono output"
+  ,"pico.video.sharpness": "Sharpness"
+  ,"pico.video.volume": "Volume %"
+  ,"pico.config.save_global": "Save global config"
+  ,"pico.config.save_game": "Save game config"
+  ,"pico.config.delete_game": "Delete game config"
+  ,"pico.config.restore_defaults": "Restore defaults"
+  ,"pico.config.loaded_game": "Loaded: game config"
+  ,"pico.config.loaded_global": "Loaded: global config"
+  ,"pico.legend.back_ok": "B-BACK   A-OKAY"
 }
