+++ lang/builtin/pl_PL.json
@@ -77,4 +77,36 @@
   ,"app.image_viewer": "Obrazy"
   ,"app.videos": "Wideo"
   ,"app.music_mp3": "Muzyka / MP3"
+  ,"pico.menu.resume": "Wznów grę"
+  ,"pico.menu.save_state": "Zapisz stan"
+  ,"pico.menu.load_state": "Wczytaj stan"
+  ,"pico.menu.disc_control": "Sterowanie dyskiem"
+  ,"pico.menu.cheats": "Kody"
+  ,"pico.menu.options": "Opcje"
+  ,"pico.menu.reset_game": "Zresetuj grę"
+  ,"pico.menu.load_new_game": "Wczytaj nową grę"
+  ,"pico.menu.exit": "Wyjdź"
+  ,"pico.options.audio_video": "Dźwięk i obraz"
+  ,"pico.options.emulator": "Opcje emulatora"
+  ,"pico.options.player_controls": "Sterowanie gracza"
+  ,"pico.options.emulator_controls": "Sterowanie emulatorem"
+  ,"pico.options.save_config": "Zapisz konfigurację"
+  ,"pico.video.show_fps": "Pokaż FPS"
+  ,"pico.video.show_cpu": "Pokaż użycie CPU %"
+  ,"pico.video.fast_forward": "Przewijanie (SELECT+R1)"
+  ,"pico.video.rewind": "Cofanie (przytrzymaj SELECT+B)"
+  ,"pico.video.aspect_ratio": "Proporcje ekranu"
+  ,"pico.video.filter": "Filtr"
+  ,"pico.video.optimize_text": "Optymalizuj tekst"
+  ,"pico.video.audio_buffer": "Bufor dźwięku"
+  ,"pico.video.mono_output": "Dźwięk mono"
+  ,"pico.video.sharpness": "Ostrość"
+  ,"pico.video.volume": "Głośność %"
+  ,"pico.config.save_global": "Zapisz konfigurację globalną"
+  ,"pico.config.save_game": "Zapisz konfigurację gry"
+  ,"pico.config.delete_game": "Usuń konfigurację gry"
+  ,"pico.config.restore_defaults": "Przywróć domyślne"
+  ,"pico.config.loaded_game": "Wczytano konfigurację gry"
+  ,"pico.config.loaded_global": "Wczytano konfigurację globalną"
+  ,"pico.legend.back_ok": "B-WSTECZ   A-OK"
 }
